Set GCC environment to freestanding
This tells the compiler to not assume it's going to link with the
standard library. The definition is not very clear, however as gcc(1)
puts it: it allows the code to run on bare metal (like a kernel).

I noticed the compiler would stop complaining about conflicting types
(-Wbuiltin-declaration-mismatch) when writing my own `putchar(char)`.