Switch back to assembly code
Apparently GCC is not able to compile for 16-bit Real Mode environments,
and because I wasn't able to run the code using BCC (Bruce's C Compiler)
I decided to switch back to assembly.

The best thing to do would be to switch to Protected Mode. That however
would prevent me from using the comfy BIOS interrupts, making this mega
complicated, and I'm definitely not ready for that yet.